int for time instead of float

+++ src/scheduler.cpp
@@ -2,19 +2,19 @@
 #include "actor.hpp"
 
 void Scheduler::updateNextActor() {
-	std::pair<float, Actor*>& activeActor = actorsQueue.at(0);
-	float activeActorTUNA = activeActor.first;
+	std::pair<int, Actor*>& activeActor = actorsQueue.at(0);
+	int activeActorTUNA = activeActor.first;
 
-	for(std::pair<float, Actor*>& a : actorsQueue) {
+	for(std::pair<int, Actor*>& a : actorsQueue) {
 		a.first -= activeActorTUNA;
 	}
 	time += activeActorTUNA;
 
-	float elapsedTime = activeActor.second->update();
+	int elapsedTime = activeActor.second->update();
 	activeActor.first += elapsedTime;
 	std::sort(actorsQueue.begin(), actorsQueue.end()); // TODO inefficient
 }
 
 void Scheduler::insertActor(Actor* actor) {
-	if(actor->ai) actorsQueue.push_back(std::pair<float, Actor*>(100, actor));
+	if(actor->ai) actorsQueue.push_back(std::pair<int, Actor*>(100, actor));
 }
